+++ package.json
@@ -4,29 +4,38 @@
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
+    "ts:check": "tsc --noEmit",
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web"
   },
   "dependencies": {
+    "@expo/webpack-config": "^19.0.0",
     "@react-native-community/masked-view": "^0.1.11",
     "@react-navigation/native": "^6.1.7",
     "expo": "~49.0.11",
+    "expo-font": "~11.4.0",
+    "expo-splash-screen": "~0.20.5",
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-native": "0.72.4",
     "react-native-paper": "^5.10.5",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
-    "react-navigation": "^4.4.4",
-    "expo-font": "~11.4.0",
-    "expo-splash-screen": "~0.20.5",
-    "react-dom": "18.2.0",
     "react-native-web": "~0.19.6",
-    "@expo/webpack-config": "^19.0.0"
+    "react-navigation": "^4.4.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0"
+    "@babel/core": "^7.20.0",
+    "@babel/plugin-transform-typescript": "^7.22.15",
+    "@tsconfig/react-native": "^3.0.2",
+    "@types/jest": "^29.5.5",
+    "@types/react": "^18.2.22",
+    "@types/react-native-vector-icons": "^6.4.14",
+    "@types/react-test-renderer": "^18.0.2",
+    "babel-plugin-module-resolver": "^5.0.0",
+    "typescript": "^5.2.2"
   },
   "private": true
 }
